refactor: 분석 채팅 구조 분리
