+++ apps/fe/src/app/(main)/analysis/[id]/page.tsx
@@ -55,7 +55,7 @@ export default function AnalysisChatPage({
       right={
         <div className="flex h-full min-h-0 flex-col">
           <div className="scrollbar-hide flex min-h-0 flex-1 flex-col gap-24 overflow-y-auto px-24 pt-32 pb-24">
-            <AnalysisChatMessageList chat={chat} />
+            <AnalysisChatMessageList viewModel={chat.messageList} />
           </div>
 
           <div className="shrink-0 px-24 pt-12 pb-24">
